feat: extract gateway_posts package (service + schemas)
Move PostsService and PostResponse/PostsPageResponse/FullPostResponse
into a new gateway_core-only gateway_posts package. get_comments_for_post
is dropped from gateway_posts.service.PostsService (moves to
gateway_comments in a later task); gateway_oss/services/posts.py now
subclasses it to keep that method available for the OSS posts API route
and MCP tool until the comments extraction lands. get_notes_for_profile
stays duplicated per the fleet-split plan.
