+++ packages/gateway_oss/pyproject.toml
@@ -16,6 +16,7 @@ dependencies = [
   "gateway_core",
   "gateway_following",
   "gateway_notes",
+  "gateway_posts",
   "gateway_notes_rest",
   "gateway_notes_mcp",
   "gateway_following_rest",
@@ -35,6 +36,7 @@ packages = ["src/gateway_oss"]
 gateway_core = { workspace = true }
 gateway_following = { workspace = true }
 gateway_notes = { workspace = true }
+gateway_posts = { workspace = true }
 gateway_notes_rest = { workspace = true }
 gateway_notes_mcp = { workspace = true }
 gateway_following_rest = { workspace = true }
