+++ pyproject.toml
@@ -27,6 +27,7 @@ gateway_oss = { workspace = true }
 gateway_core = { workspace = true }
 gateway_following = { workspace = true }
 gateway_notes = { workspace = true }
+gateway_posts = { workspace = true }
 gateway_notes_rest = { workspace = true }
 gateway_notes_mcp = { workspace = true }
 gateway_following_rest = { workspace = true }
@@ -44,7 +45,7 @@ select = ["E", "F", "I", "UP"]
 ignore = ["E501"]
 
 [tool.ruff.lint.isort]
-known-first-party = ["gateway_core", "gateway_following", "gateway_notes", "gateway_notes_rest", "gateway_notes_mcp", "gateway_following_rest", "gateway_following_mcp", "gateway_oss", "gateway_rest_common", "gateway_mcp_common", "substack_gateway"]
+known-first-party = ["gateway_core", "gateway_following", "gateway_notes", "gateway_notes_rest", "gateway_notes_mcp", "gateway_following_rest", "gateway_following_mcp", "gateway_posts", "gateway_oss", "gateway_rest_common", "gateway_mcp_common", "substack_gateway"]
 
 [tool.ruff.format]
 quote-style = "double"
@@ -63,6 +64,7 @@ testpaths = [
   "packages/gateway_core/tests",
   "packages/gateway_following/tests",
   "packages/gateway_notes/tests",
+  "packages/gateway_posts/tests",
 ]
 addopts = "--import-mode=importlib"
